+++ Modules/Intent.Modules.Application.DomainInteractions/DomainInteractionsManager.cs
@@ -216,9 +216,9 @@ public class DomainInteractionsManager
                 {
                     //var expression = CreateQueryFilterExpression(queryMapping, out var requiredStatements);
 
-                    if (TryGetPaginationValues(associationEnd, _csharpMapping, out var pageNo, out var pageSize, out var orderBy))
+                    if (TryGetPaginationValues(associationEnd, _csharpMapping, out var pageNo, out var pageSize, out var orderBy, out var orderByIsNUllable))
                     {
-                        queryInvocation = dataAccess.FindAllAsync(queryMapping, pageNo, pageSize, orderBy, out var requiredStatements);
+                        queryInvocation = dataAccess.FindAllAsync(queryMapping, pageNo, pageSize, orderBy, orderByIsNUllable, out var requiredStatements);
                         prerequisiteStatement.AddRange(requiredStatements);
                     }
                     else if (associationEnd.TypeReference.IsCollection)
@@ -330,7 +330,7 @@ public class DomainInteractionsManager
         //This is being done for Dapper
         bool hasUnitOfWork = _template.TryGetTemplate<ITemplate>(TemplateRoles.Domain.UnitOfWork, out _);
 
-		dataAccessProvider = new RepositoryDataAccessProvider(InjectService(repositoryInterface, handlerClass), _template, _csharpMapping, hasUnitOfWork);
+		dataAccessProvider = new RepositoryDataAccessProvider(InjectService(repositoryInterface, handlerClass), _template, _csharpMapping, hasUnitOfWork, foundEntity);
         return true;
     }
 
@@ -961,8 +961,9 @@ public class DomainInteractionsManager
         return true;
     }
 
-    private bool TryGetPaginationValues(IAssociationEnd associationEnd, CSharpClassMappingManager mappingManager, out string pageNo, out string pageSize, out string? orderBy)
+    private bool TryGetPaginationValues(IAssociationEnd associationEnd, CSharpClassMappingManager mappingManager, out string pageNo, out string pageSize, out string? orderBy, out bool orderByIsNullable)
     {
+        orderByIsNullable = false;
         var handler = (IElement)associationEnd.OtherEnd().TypeReference.Element;
         var returnsPagedResult = IsResultPaginated(handler.TypeReference);
 
@@ -978,6 +979,7 @@ public class DomainInteractionsManager
         }
         else
         {
+            orderByIsNullable = orderByVar.TypeReference.IsNullable;
             orderBy = $"{(accessVariable != null ? $"{accessVariable}." : "")}{handler.ChildElements.Single(IsOrderByParam)?.Name}";
         }