+++ Modules/Intent.Modules.Application.ServiceImplementations.CRUD/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.0.12
+
+- Improvement: Support for nullable `OrderBy` on Paging.
+
 ### Version 5.0.11
 
 - Improvement: Add support for `OrderBy` on pagination.